refactor: move replica under blob package, fix ability names
- prefix ability for replica with /blob
