+++ capabilities/blob/replica/allocate_test.go
@@ -5,8 +5,8 @@ import (
 
 	"github.com/stretchr/testify/require"
 
+	"github.com/storacha/go-libstoracha/capabilities/blob/replica"
 	"github.com/storacha/go-libstoracha/capabilities/internal/testutil"
-	"github.com/storacha/go-libstoracha/capabilities/replica"
 	"github.com/storacha/go-libstoracha/capabilities/space/blob"
 )
