+++ capabilities/blob/replica/replica.ipldsch
