+++ capabilities/blob/replica/schema.go
